+++ holdfast/src/commonTest/kotlin/com/vynatix/holdfast/StoreLockTest.kt
@@ -1,3 +1,5 @@
+@file:OptIn(StoreInternalApi::class)
+
 package com.vynatix.holdfast
 
 import kotlinx.atomicfu.atomic